refactor: move space to vectors
