+++ src/coordinax/__init__.py
@@ -10,10 +10,9 @@ from .setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("coordinax", RUNTIME_TYPECHECKER):
     from . import angle, distance, frames, operators
-    from ._src import exceptions, funcs, space, typing, utils, vectors
+    from ._src import exceptions, funcs, typing, utils, vectors
     from ._src.exceptions import *
     from ._src.funcs import *
-    from ._src.space import *
     from ._src.typing import *
     from ._src.utils import *
     from ._src.vectors.base import *
@@ -22,6 +21,7 @@ with install_import_hook("coordinax", RUNTIME_TYPECHECKER):
     from ._src.vectors.d3 import *
     from ._src.vectors.d4 import *
     from ._src.vectors.dn import *
+    from ._src.vectors.space import *
     from ._version import version as __version__
     from .distance import Distance
 
@@ -42,7 +42,7 @@ __all__ += vectors.d2.__all__
 __all__ += vectors.d3.__all__
 __all__ += vectors.d4.__all__
 __all__ += vectors.dn.__all__
-__all__ += space.__all__
+__all__ += vectors.space.__all__
 __all__ += exceptions.__all__
 __all__ += typing.__all__
 __all__ += utils.__all__
@@ -51,7 +51,6 @@ __all__ += utils.__all__
 # Cleanup
 del (
     vectors,
-    space,
     exceptions,
     transform,
     typing,
