+++ src/coordinax/_src/vectors/space/compat.py
